// dashboard/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        RunSummaryResponse: {
            period: components["schemas"]["TimeRangePeriod"] | null;
            /** Interval */
            interval: ("minute" | "hour" | "day") | null;
            /** Points */
            points: components["schemas"]["TrendPoint"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}